Rollback Stage updation

+++ Jenkinsfile
@@ -1,4 +1,5 @@
 pipeline {
+
     agent any
 
     environment {
@@ -7,6 +8,7 @@ pipeline {
     }
 
     stages {
+
         stage('Clean Workspace') {
             steps {
                 cleanWs()
@@ -20,11 +22,36 @@ pipeline {
             }
         }
 
+        stage('Python Validation') {
+            steps {
+                sh '''
+                python3 -m py_compile app.py
+                '''
+            }
+        }
+
         stage('Build Docker Image') {
             steps {
                 sh '''
-                docker build -t task-manager-app:${BUILD_NUMBER} .
-                docker tag task-manager-app:${BUILD_NUMBER} task-manager-app:latest
+                docker build -t ${IMAGE_NAME}:${BUILD_NUMBER} .
+                docker tag ${IMAGE_NAME}:${BUILD_NUMBER} ${IMAGE_NAME}:latest
+                '''
+            }
+        }
+
+        stage('Trivy Security Scan') {
+            steps {
+                sh '''
+                trivy image --severity HIGH,CRITICAL ${IMAGE_NAME}:latest || true
+                '''
+            }
+        }
+
+        stage('Backup Current Image') {
+            steps {
+                sh '''
+                docker image inspect ${IMAGE_NAME}:latest >/dev/null 2>&1 && \
+                docker tag ${IMAGE_NAME}:latest ${IMAGE_NAME}:previous || true
                 '''
             }
         }
@@ -32,43 +59,81 @@ pipeline {
         stage('Deploy Application') {
             steps {
                 sh '''
-                docker stop task-manager-app || true
-                docker rm task-manager-app || true
+                docker stop ${APP_NAME} || true
+                docker rm ${APP_NAME} || true
 
                 docker run -d \
-                  --name task-manager-app \
+                  --name ${APP_NAME} \
                   --env-file /opt/task-manager/.env \
                   -p 5000:5000 \
-                  task-manager-app:latest
+                  ${IMAGE_NAME}:latest
                 '''
             }
         }
 
         stage('Health Check') {
             steps {
+                script {
+
+                    def healthStatus = sh(
+                        script: '''
+                        sleep 20
+
+                        for i in 1 2 3
+                        do
+                            curl -f http://localhost:5000/health && exit 0
+                            echo "Health check failed. Retrying..."
+                            sleep 10
+                        done
+
+                        exit 1
+                        ''',
+                        returnStatus: true
+                    )
+
+                    if (healthStatus != 0) {
+
+                        echo "Deployment failed. Starting rollback..."
+
+                        sh '''
+                        docker stop task-manager-app || true
+                        docker rm task-manager-app || true
+
+                        docker run -d \
+                          --name task-manager-app \
+                          --env-file /opt/task-manager/.env \
+                          -p 5000:5000 \
+                          task-manager-app:previous
+                        '''
+
+                        error("Rollback executed due to failed health check.")
+                    }
+                }
+            }
+        }
+
+        stage('Docker Cleanup') {
+            steps {
                 sh '''
-                sleep 20
-
-                for i in 1 2 3
-                do
-                    curl -f http://localhost:5000/health && exit 0
-                    echo "Health check failed. Retrying..."
-                    sleep 10
-                done
-
-                exit 1
+                docker image prune -f
+                docker container prune -f
                 '''
             }
         }
     }
 
     post {
+
         success {
             echo 'Deployment Successful'
         }
 
         failure {
-            echo 'Deployment Failed'
+            echo 'Deployment Failed and Rollback Executed'
+        }
+
+        always {
+            cleanWs()
         }
     }
 }
